cr: mount code about cr in a separate file
